Save pending families to the REST API

+++ app/src/main/java/org/fundacionparaguaya/advisorapp/data/remote/intermediaterepresentation/FamilyIr.java
@@ -2,15 +2,13 @@ package org.fundacionparaguaya.advisorapp.data.remote.intermediaterepresentation
 
 import com.google.gson.annotations.SerializedName;
 
-import org.fundacionparaguaya.advisorapp.models.Family;
-
 /**
  * The intermediate representation of the family from the remote database.
  */
 
 public class FamilyIr {
     @SerializedName("familyId")
-    int id;
+    long id;
     @SerializedName("code")
     String code;
     @SerializedName("name")
